Just moved the basic program into a headeronly library
